+++ bsnes/snes/interface/interface.hpp
@@ -1,6 +1,6 @@
 class Interface {
 public:
-  virtual void videoRefresh(const uint16_t *data, bool hires, bool interlace, bool overscan);
+  virtual void videoRefresh(const uint32_t *data, bool hires, bool interlace, bool overscan);
   virtual void audioSample(int16_t lsample, int16_t rsample);
   virtual int16_t inputPoll(bool port, Input::Device device, unsigned index, unsigned id);
