+++ bsnes/snes/ppu/ppu.hpp
@@ -23,8 +23,8 @@ public:
   ~PPU();
 
 private:
-  uint16 *surface;
-  uint16 *output;
+  uint32 *surface;
+  uint32 *output;
 
   uint8 ppu1_version;
   uint8 ppu2_version;
